move image data selector to slice
