+++ app/javascript/mastodon/features/account_edit/modals/image_delete.tsx
@@ -1,8 +1,9 @@
 import type { FC } from 'react';
 
+import type { ImageLocation } from '@/mastodon/reducers/slices/profile_edit';
+
 import { DialogModal } from '../../ui/components/dialog_modal';
 import type { DialogModalProps } from '../../ui/components/dialog_modal';
-import type { ImageLocation } from '../components/image_edit';
 
 export const ImageDeleteModal: FC<
   DialogModalProps & { location: ImageLocation }
