+++ app/javascript/mastodon/features/account_edit/modals/image_upload.tsx
@@ -1,11 +1,40 @@
+import { useState } from 'react';
 import type { FC } from 'react';
+
+import { FormattedMessage } from 'react-intl';
+
+import { selectImageInfo } from '@/mastodon/reducers/slices/profile_edit';
+import type { ImageLocation } from '@/mastodon/reducers/slices/profile_edit';
+import { useAppSelector } from '@/mastodon/store';
 
 import { DialogModal } from '../../ui/components/dialog_modal';
 import type { DialogModalProps } from '../../ui/components/dialog_modal';
-import type { ImageLocation } from '../components/image_edit';
 
 export const ImageUploadModal: FC<
   DialogModalProps & { location: ImageLocation }
-> = ({ onClose }) => {
-  return <DialogModal title='TODO' onClose={onClose} />;
+> = ({ onClose, location }) => {
+  const { src } = useAppSelector((state) => selectImageInfo(state, location));
+  const [step] = useState<'select' | 'crop' | 'alt'>('select');
+
+  return (
+    <DialogModal
+      title={
+        src ? (
+          <FormattedMessage
+            id='account_edit.upload_modal.title_replace'
+            defaultMessage='Replace profile photo'
+          />
+        ) : (
+          <FormattedMessage
+            id='account_edit.upload_modal.title_add'
+            defaultMessage='Add profile photo'
+          />
+        )
+      }
+      onClose={onClose}
+      noCancelButton
+    >
+      {step}
+    </DialogModal>
+  );
 };
